+++ CHANGELOG.md
@@ -48,7 +48,7 @@
 - Change constructor of `Pim\Bundle\BaseConnectorBundle\Processor\TransformerProcessor`. `Doctrine\Common\Persistence\ManagerRegistry` is used as fourth argument and is mandatory now. The data class is the fifth argument.
 - Change constructor of `Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter\*` and `Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter\*` to remove the CatalogContext
 - Remove ProductRepositoryInterface::findOneBy (still a native support for ORM)
-- Add ProductRepositoryInterface::findOneByIdentifier
+- Add ProductRepositoryInterface::findOneByIdentifier and findOneById
 - Remove ProductRepositoryInterface::buildByScope
 - Remove ProductRepositoryInterface::findByExistingFamily
 - Remove ProductRepositoryInterface::findAllByAttributes